pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.6",
    "mcp-use>=1.5.1",
    "playwright>=1.57.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
]
